scm-core => add camel thread config and personId in login time

+++ scm-uaa-common/src/main/java/ir/daneshrefah/scm/uaa/common/model/person/GeneralPerson.java
@@ -13,7 +13,7 @@ import lombok.Data;
  * @since 2023-07-19
  */
 @Data
-public abstract class GeneralPerson extends BaseModel<Integer> {
+public abstract class GeneralPerson extends BaseModel<Long> {
 
     private String username;
     private Boolean active;
